Switch from packer.nvim to lazy.nvim
Use lazy.nvim -- supposedly faster. Plugins setup now done using lazy's
API. Removed setting of color from options to plugins file. Having it in
multiple places is confusing -- and it had to be set as one of the first
things for lazy.nvim to work. Remove spotify rule which places spotify
on the 10th workspace, but always in a maximized way which is annoying.